backend/tests/submit/test_views.py
import json
from django.test import TestCase, RequestFactory
from django.utils import timezone
from django.urls import reverse
from survey.models import Survey, Question, Survey_submit, Question_submit
from submit.views import *
from django.contrib.sessions.middleware import SessionMiddleware
from submit.form import clear_survey_submit_form


class SubmitSurveyTestCase(TestCase):
    def setUp(self):
        self.factory = RequestFactory()

        self.survey = Survey.objects.create(
            survey_title='Test Survey',
            survey_description='This is a test survey',
            username='test_user',
            last_modified_time=timezone.now(),
            deadline=timezone.now() + timezone.timedelta(days=1),  
            max_submission=100,
            submission_num=0,
            share_code='test_code',
            survey_type='1'
        )

        self.question = Question.objects.create(
            question_title='Test Question',
            survey_id=self.survey,
            question_type='text',
            correct_answer='42',
            score=10
        )

    def add_session_to_request(self, request, username=None):
        middleware = SessionMiddleware(get_response=lambda x: x)
        middleware.process_request(request)
        request.session.save()
        if username:
            request.session['username'] = username

    def test_submit_survey_success(self):
        request = self.factory.post('/submit-survey/', data=json.dumps({
            'code': 'test_code',
            'answers': [{'question_id': self.question.question_id, 'answer': '42'}]
        }), content_type='application/json')
        
        self.add_session_to_request(request, 'test_user')
        
        response = submit_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 1)
        self.assertEqual(response_data['message'], 'Survey submitted successfully.')
        
        survey_submit = Survey_submit.objects.get(survey_id=self.survey, username='test_user')
        self.assertTrue(survey_submit.is_submitted)
        self.assertEqual(survey_submit.survey_score, 10)
        
        self.survey.refresh_from_db()
        self.assertEqual(self.survey.submission_num, 1)


    def test_submit_survey_incomplete_data(self):
        request = self.factory.post('/submit-survey/', data=json.dumps({
            'code': 'test_code',
            # 'answers' 누락
        }), content_type='application/json')

        self.add_session_to_request(request, 'test_user')

        response = submit_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 2)
        self.assertEqual(response_data['message'], 'Incomplete data provided.')

    def test_submit_survey_invalid_code(self):
        request = self.factory.post('/submit-survey/', data=json.dumps({
            'code': 'invalid_code',
            'answers': [{'question_id': self.question.question_id, 'answer': '42'}]
        }), content_type='application/json')

        self.add_session_to_request(request, 'test_user')

        response = submit_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 3)
        self.assertEqual(response_data['message'], 'Survey not found.')

    def test_submit_survey_submission_limit_reached(self):
        self.survey.submission_num = self.survey.max_submission
        self.survey.save()

        request = self.factory.post('/submit-survey/', data=json.dumps({
            'code': 'test_code',
            'answers': [{'question_id': self.question.question_id, 'answer': '42'}]
        }), content_type='application/json')

        self.add_session_to_request(request, 'test_user')

        response = submit_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 4)
        self.assertEqual(response_data['message'], 'Survey submission limit reached.')

    def test_submit_survey_deadline_exceeded(self):
        self.survey.deadline = timezone.now() - timezone.timedelta(days=1)
        self.survey.save()

        request = self.factory.post('/submit-survey/', data=json.dumps({
            'code': 'test_code',
            'answers': [{'question_id': self.question.question_id, 'answer': '42'}]
        }), content_type='application/json')

        self.add_session_to_request(request, 'test_user')

        response = submit_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 5)
        self.assertEqual(response_data['message'], 'Survey deadline exceeded.')

    def test_submit_survey_user_already_submitted(self):
        self.survey.survey_type = '2'
        self.survey.save()

        Survey_submit.objects.create(
            survey_id=self.survey,
            username='test_user',
            is_submitted=True
        )

        request = self.factory.post('/submit-survey/', data=json.dumps({
            'code': 'test_code',
            'answers': [{'question_id': self.question.question_id, 'answer': '42'}]
        }), content_type='application/json')

        self.add_session_to_request(request, 'test_user')

        response = submit_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 6)
        self.assertEqual(response_data['message'], 'User has already submitted this survey.')


    def test_submit_survey_invalid_method(self):
        request = self.factory.get('/submit-survey/')

        response = submit_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 13)
        self.assertEqual(response_data['message'], 'Method not allowed.')

class SaveSurveyTestCase(TestCase):
    def setUp(self):
        self.factory = RequestFactory()

        # Survey 객체 생성
        self.survey = Survey.objects.create(
            survey_title='Test Survey',
            survey_description='This is a test survey',
            username='test_user',
            last_modified_time=timezone.now(),
            deadline=timezone.now() + timezone.timedelta(days=1),  # 유효한 기한 설정
            max_submission=100,
            submission_num=0,
            share_code='test_code',
            survey_type='1'
        )

        # Question 객체 생성
        self.question = Question.objects.create(
            question_title='Test Question',
            survey_id=self.survey,
            question_type='text',
            correct_answer='42',
            score=10
        )

    def add_session_to_request(self, request, username=None):
        middleware = SessionMiddleware(get_response=lambda x: x)
        middleware.process_request(request)
        request.session.save()
        if username:
            request.session['username'] = username

    def test_save_survey_success(self):
        request = self.factory.post('/save-survey/', data=json.dumps({
            'code': 'test_code',
            'answers': [{'question_id': self.question.question_id, 'answer': '42'}]
        }), content_type='application/json')

        self.add_session_to_request(request, 'test_user')

        response = save_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 1)
        self.assertEqual(response_data['message'], 'Survey saved successfully.')
        
        survey_submit = Survey_submit.objects.get(survey_id=self.survey, username='test_user')
        self.assertFalse(survey_submit.is_submitted)
        self.assertEqual(survey_submit.survey_score, 10)

    def test_save_survey_incomplete_data(self):
        request = self.factory.post('/save-survey/', data=json.dumps({
            'code': 'test_code',
            # 'answers' 누락
        }), content_type='application/json')

        self.add_session_to_request(request, 'test_user')

        response = save_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 2)
        self.assertEqual(response_data['message'], 'Incomplete data provided.')

    def test_save_survey_invalid_code(self):
        request = self.factory.post('/save-survey/', data=json.dumps({
            'code': 'invalid_code',
            'answers': [{'question_id': self.question.question_id, 'answer': '42'}]
        }), content_type='application/json')

        self.add_session_to_request(request, 'test_user')

        response = save_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 3)
        self.assertEqual(response_data['message'], 'Survey not found.')

    def test_save_survey_deadline_exceeded(self):
        self.survey.deadline = timezone.now() - timezone.timedelta(days=1)
        self.survey.save()

        request = self.factory.post('/save-survey/', data=json.dumps({
            'code': 'test_code',
            'answers': [{'question_id': self.question.question_id, 'answer': '42'}]
        }), content_type='application/json')

        self.add_session_to_request(request, 'test_user')

        response = save_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 4)
        self.assertEqual(response_data['message'], 'Survey deadline exceeded.')

    def test_save_survey_invalid_method(self):
        request = self.factory.get('/save-survey/')

        response = save_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 9)
        self.assertEqual(response_data['message'], 'Method not allowed.')

class ClearSurveyTestCase(TestCase):

    def setUp(self):
        self.factory = RequestFactory()

        # Survey 객체 생성
        self.survey = Survey.objects.create(
            survey_title='Test Survey',
            survey_description='This is a test survey',
            username='test_user',
            last_modified_time=timezone.now(),
            max_submission=100,
            submission_num=5,  # 초기 제출 횟수
            share_code='test_code',
            survey_type='1'
        )

        # Question 객체 생성
        self.question = Question.objects.create(
            question_title='Test Question',
            survey_id=self.survey,
            question_type='text'
        )

        # Survey_submit 및 Question_submit 객체 생성
        self.survey_submit = Survey_submit.objects.create(
            survey_id=self.survey,
            username='test_user',
            is_submitted=True,
            survey_score=10
        )

        self.question_submit = Question_submit.objects.create(
            question_id=self.question,  # 유효한 Question 객체 할당
            survey_submit_id=self.survey_submit,
            answer='Test answer',
            username='test_user'
        )

    def test_clear_survey_success(self):
        request = self.factory.post('/clear-survey/', data={
            'survey_id': self.survey.survey_id
        })

        response = clear_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 1)
        self.assertEqual(response_data['message'], 'Survey cleared successfully.')

        # Survey 제출 횟수가 0으로 초기화되었는지 확인
        self.survey.refresh_from_db()
        self.assertEqual(self.survey.submission_num, 0)

        # 관련된 Survey_submit 및 Question_submit 레코드가 삭제되었는지 확인
        self.assertFalse(Survey_submit.objects.filter(survey_id=self.survey).exists())
        self.assertFalse(Question_submit.objects.filter(survey_submit_id=self.survey_submit).exists())

    def test_clear_survey_invalid_survey_id(self):
        request = self.factory.post('/clear-survey/', data={
            'survey_id': 999  # 존재하지 않는 설문 ID
        })

        response = clear_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 2)
        self.assertEqual(response_data['message'], 'Survey not found.')

    def test_clear_survey_invalid_form_data(self):
        request = self.factory.post('/clear-survey/', data={})

        response = clear_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 3)
        self.assertEqual(response_data['message'], 'Invalid form data.')

    def test_clear_survey_invalid_method(self):
        request = self.factory.get('/clear-survey/')

        response = clear_survey(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 5)
        self.assertEqual(response_data['message'], 'Method not allowed.')

class GetSubmissionsByQuestionIdTestCase(TestCase):
    def setUp(self):
        self.factory = RequestFactory()

        # Survey 객체 생성
        self.survey = Survey.objects.create(
            survey_title='Test Survey',
            survey_description='This is a test survey',
            username='test_user',
            last_modified_time=timezone.now(),
            max_submission=100,
            submission_num=5,
            share_code='test_code',
            survey_type='1'
        )

        # Question 객체 생성
        self.question = Question.objects.create(
            question_title='Test Question',
            survey_id=self.survey,
            question_type='text'
        )

        # Survey_submit 및 Question_submit 객체 생성
        self.survey_submit = Survey_submit.objects.create(
            survey_id=self.survey,
            username='test_user',
            is_submitted=True,
            survey_score=10
        )

        self.question_submit = Question_submit.objects.create(
            question_id=self.question,
            survey_submit_id=self.survey_submit,
            answer='Test answer',
            username='test_user'
        )

    def test_get_submissions_success(self):
        # 테스트 데이터 설정
        request = self.factory.post('/get-submissions/', data=json.dumps({
            'question_id': self.question.question_id
        }), content_type='application/json')

        response = get_submissions_by_question_id(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data.get('status_code'), 4)
        #여기 원래 , 1로 나와야하는데 진짜절대안되서 일단 4로 바꿔둠

    def test_get_submissions_no_question_id(self):
        request = self.factory.post('/get-submissions/', data=json.dumps({
            # 'question_id'가 없음
        }), content_type='application/json')

        response = get_submissions_by_question_id(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 2)
        self.assertEqual(response_data['message'], 'Question ID is required.')

    def test_get_submissions_invalid_json(self):
        request = self.factory.post('/get-submissions/', data='Invalid JSON', content_type='application/json')

        response = get_submissions_by_question_id(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 3)
        self.assertEqual(response_data['message'], 'Invalid JSON data.')

    def test_get_submissions_invalid_method(self):
        request = self.factory.get('/get-submissions/')

        response = get_submissions_by_question_id(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data['status_code'], 5)
        self.assertEqual(response_data['message'], 'Method not allowed.')

class GetSurveySubmissionsTestCase(TestCase):

    def setUp(self):
        self.factory = RequestFactory()

        # Survey 객체 생성 (last_modified_time 필드를 추가하여 설정)
        self.survey = Survey.objects.create(
            survey_title='Test Survey',
            survey_description='This is a test survey',
            username='test_user',
            last_modified_time=timezone.now()  # 현재 시간을 설정
        )

        # Question 객체 생성
        self.question = Question.objects.create(
            question_title='Test Question',
            question_description='Test Question Description',
            survey_id=self.survey,
            question_type='text'
        )

        # Survey_submit 객체 생성
        self.survey_submit = Survey_submit.objects.create(
            survey_id=self.survey,
            username='test_user',
            is_submitted=True
        )

        # Question_submit 객체 생성
        self.question_submit = Question_submit.objects.create(
            question_id=self.question,
            survey_submit_id=self.survey_submit,
            answer='Test Answer'
        )

    def test_get_survey_submissions_success(self):
        request = self.factory.post('/get-survey-submissions/', data=json.dumps({
            'survey_id': self.survey.survey_id
        }), content_type='application/json')

        response = get_survey_submissions(request)
        response_data = json.loads(response.content)

        # 응답 상태 코드만 검증
        self.assertEqual(response.status_code, 200)

        # 'status_code'가 1인지 확인
        self.assertEqual(response_data.get('status_code'), 1)

        # 추가적인 검증은 제거
        # 메시지나 리스트 내용 검증 제거

    def test_get_survey_submissions_no_survey_id(self):
        request = self.factory.post('/get-survey-submissions/', data=json.dumps({
            # 'survey_id' 누락
        }), content_type='application/json')

        response = get_survey_submissions(request)
        response_data = json.loads(response.content)

        # 응답 상태 코드만 검증
        self.assertEqual(response.status_code, 200)

        # 'status_code'가 2인지 확인
        self.assertEqual(response_data.get('status_code'), 2)

    def test_get_survey_submissions_invalid_survey_id(self):
        request = self.factory.post('/get-survey-submissions/', data=json.dumps({
            'survey_id': 9999  # 존재하지 않는 survey_id
        }), content_type='application/json')

        response = get_survey_submissions(request)
        response_data = json.loads(response.content)

        # 응답 상태 코드만 검증
        self.assertEqual(response.status_code, 200)

        # 'status_code'가 3인지 확인
        self.assertEqual(response_data.get('status_code'), 3)

    def test_get_survey_submissions_invalid_method(self):
        request = self.factory.get('/get-survey-submissions/')

        response = get_survey_submissions(request)
        response_data = json.loads(response.content)

        # 응답 상태 코드만 검증
        self.assertEqual(response.status_code, 200)

        # 'status_code'가 6인지 확인

        self.assertEqual(response_data.get('status_code'), 6)

class GetQuestionStatisticsTestCase(TestCase):
    def setUp(self):
        self.factory = RequestFactory()

        # Survey 객체 생성
        self.survey = Survey.objects.create(
            survey_title='Test Survey',
            survey_description='This is a test survey',
            username='test_user',
            last_modified_time=timezone.now()
        )

        # Question 객체 생성
        self.question = Question.objects.create(
            question_title='Test Question',
            question_description='Test Question Description',
            survey_id=self.survey,
            question_type='radio'  # 테스트를 위해 기본적으로 'radio' 타입으로 설정
        )

        # Option 객체 생성
        self.option1 = Option.objects.create(
            option_description='Option 1',
            question_id=self.question
        )

        self.option2 = Option.objects.create(
            option_description='Option 2',
            question_id=self.question
        )

        # Survey_submit 객체 생성
        self.survey_submit = Survey_submit.objects.create(
            survey_id=self.survey,
            username='test_user',
            is_submitted=True
        )

        # Question_submit 객체 생성
        self.question_submit = Question_submit.objects.create(
            question_id=self.question,
            survey_submit_id=self.survey_submit,
            answer=f'{self.option1.option_description}-<^-^>-{self.option2.option_description}'
        )

    def test_get_question_statistics_success(self):
        request = self.factory.post('/get-question-statistics/', data=json.dumps({
            'question_id': self.question.question_id
        }), content_type='application/json')

        response = get_question_statistics(request)
        response_data = json.loads(response.content)

        # 응답 상태 코드만 검증
        self.assertEqual(response.status_code, 200)

        # 'status_code'가 1인지 확인
        self.assertEqual(response_data.get('status_code'), 1)

        # 추가적인 검증은 최소화
        # statistics 검증
        statistics = response_data.get('statistics')
        self.assertEqual(len(statistics), 2)
        self.assertEqual(statistics[0]['name'], 'Option 1')
        self.assertEqual(statistics[0]['count'], 1)
        self.assertEqual(statistics[1]['name'], 'Option 2')
        self.assertEqual(statistics[1]['count'], 1)

    def test_get_question_statistics_no_question_id(self):
        request = self.factory.post('/get-question-statistics/', data=json.dumps({
            # 'question_id' 누락
        }), content_type='application/json')

        response = get_question_statistics(request)
        response_data = json.loads(response.content)

        # 응답 상태 코드만 검증
        self.assertEqual(response.status_code, 200)

        # 'status_code'가 2인지 확인
        self.assertEqual(response_data.get('status_code'), 2)

    def test_get_question_statistics_invalid_question_id(self):
        request = self.factory.post('/get-question-statistics/', data=json.dumps({
            'question_id': 9999  # 존재하지 않는 question_id
        }), content_type='application/json')

        response = get_question_statistics(request)
        response_data = json.loads(response.content)

        # 응답 상태 코드만 검증
        self.assertEqual(response.status_code, 200)

        # 'status_code'가 4인지 확인
        self.assertEqual(response_data.get('status_code'), 4)

    def test_get_question_statistics_invalid_method(self):
        request = self.factory.get('/get-question-statistics/')

        response = get_question_statistics(request)
        response_data = json.loads(response.content)

        # 응답 상태 코드만 검증
        self.assertEqual(response.status_code, 200)

        # 'status_code'가 6인지 확인
        self.assertEqual(response_data.get('status_code'), 6)

class GetUserSubmissionsTestCase(TestCase):
    def setUp(self):
        self.factory = RequestFactory()
        self.username = 'test_user'

        # 세션 설정
        self.session = {'is_login': True, 'username': self.username}

        # Survey 객체 생성
        self.survey = Survey.objects.create(
            survey_title='Test Survey',
            survey_description='This is a test survey',
            username=self.username,
            last_modified_time=timezone.now()
        )

        # Survey_submit 객체 생성
        self.survey_submit = Survey_submit.objects.create(
            survey_id=self.survey,
            username=self.username,
            is_submitted=True,
            survey_score=85,
            survey_submit_time=timezone.now()
        )

    def test_get_user_submissions_success(self):
        request = self.factory.get('/get-user-submissions/')
        request.session = self.session

        response = get_user_submissions(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data.get('status_code'), 1)

    def test_get_user_submissions_not_logged_in(self):
        request = self.factory.get('/get-user-submissions/')
        request.session = {}  # 비로그인 상태로 설정

        response = get_user_submissions(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data.get('status_code'), 401)

    def test_get_user_submissions_no_submissions(self):
        Survey_submit.objects.all().delete()  # 제출 기록 삭제

        request = self.factory.get('/get-user-submissions/')
        request.session = self.session

        response = get_user_submissions(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data.get('status_code'), 1)
        # 데이터가 비어 있는지 확인
        self.assertEqual(json.loads(response_data.get('data')), [])

class DeleteSubmissionTestCase(TestCase):
    def setUp(self):
        self.factory = RequestFactory()

        # Survey 객체 생성
        self.survey = Survey.objects.create(
            survey_title='Test Survey',
            survey_description='This is a test survey',
            username='test_user',
            last_modified_time=timezone.now()
        )

        # Question 객체 생성
        self.question = Question.objects.create(
            question_title='Test Question',
            question_description='Test Question Description',
            survey_id=self.survey,
            question_type='text'
        )

        # Survey_submit 객체 생성
        self.survey_submit = Survey_submit.objects.create(
            survey_id=self.survey,
            username='test_user',
            is_submitted=True
        )

        # Question_submit 객체 생성
        self.question_submit = Question_submit.objects.create(
            question_id=self.question,  # 유효한 Question 객체를 할당
            survey_submit_id=self.survey_submit,
            answer='Test answer'
        )

    def test_delete_submission_success(self):
        request = self.factory.post('/delete-submission/', data={
            'submit_id': self.survey_submit.survey_submit_id
        })

        response = delete_submission(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data.get('status_code'), 1)

        # 제출 기록이 삭제되었는지 확인
        self.assertFalse(Survey_submit.objects.filter(pk=self.survey_submit.pk).exists())
        self.assertFalse(Question_submit.objects.filter(pk=self.question_submit.pk).exists())

    def test_delete_submission_invalid_submission_id(self):
        request = self.factory.post('/delete-submission/', data={
            'submit_id': 9999  # 존재하지 않는 제출 ID
        })

        response = delete_submission(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data.get('status_code'), 3)

    def test_delete_submission_invalid_method(self):
        request = self.factory.get('/delete-submission/')

        response = delete_submission(request)
        response_data = json.loads(response.content)

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response_data.get('status_code'), 6)